Add flagstat output and test empty mpileup

+++ src/code.sh
@@ -8,7 +8,7 @@ set -e -x -o pipefail
 dx-download-all-inputs --except ref_genome --parallel
 
 # make output folders
-mkdir -p ~/out/varscan_vcf/output ~/out/varscan_vcf_bed/output
+mkdir -p ~/out/varscan_vcf/output ~/out/varscan_vcf_bed/output ~/out/flagstat/QC/
 
 echo $min_coverage
 echo $min_reads2
@@ -87,18 +87,26 @@ if [ $(samtools view -c ${bam_file_path[i]}) -eq 0 ]; then
 	echo "empty BAM. skipping...."
 # if not empty perform variant calling
 else
-	# generate an mpileup from bam file then pipe to Varscan mpileupcns function
-	# write vcf direct to output folder
-	samtools mpileup -f $genome_file -B -d 500000 -q 1 ${bam_file_path[i]}| \
-	$java -jar /usr/bin/VarScan.v2.4.3.jar mpileup2cns $opts > ~/out/varscan_vcf/output/${bam_file_prefix[i]}.varscan.vcf
-	# Rename sample in vcf to corrospond to bam file name (aka sample name). Varscan defult is to name samples 'Sample1'
-	sed -i 's/Sample1/'"${bam_file_prefix[i]}"'/' ~/out/varscan_vcf/output/${bam_file_prefix[i]}.varscan.vcf
-
-	# filter vcf to disply variants located within genomic regions specified by the bed file input.
-	# output vcf direct to the output folder
-	if [ "$bed_file" != "" ]; then
-		 sed 's/chr//' ~/out/varscan_vcf/output/${bam_file_prefix[i]}.varscan.vcf > ${bam_file_prefix[i]}.temp.vcf
-	 	/usr/bin/bedtools2/bin/bedtools intersect -header -a ${bam_file_prefix[i]}.temp.vcf -b ${bed_file_path} > ~/out/varscan_vcf_bed/output/${bam_file_prefix[i]}.varscan.bedfiltered.vcf	 	
+	# generate an mpileup from bam file
+	samtools mpileup -f $genome_file -B -d 500000 -q 1 ${bam_file_path[i]} > ${bam_file_prefix[i]}.mpileup
+	# generate a flagstat output 
+	samtools flagstat  ${bam_file_path[i]} > ~/out/flagstat/QC/${bam_file_prefix[i]}.flagstat
+	# test if the mpileup file is empty - if it is skip varscan variant calling
+	if [ $(cat ${bam_file_prefix[i]}.mpileup | wc -l ) -eq 0 ]; then
+		# skip and write to stdout
+		echo "empty mpileup file. skipping...."
+	else
+		#Call varscan on mpileup file using mpileupcns function. write vcf direct to output folder
+		cat ${bam_file_prefix[i]}.mpileup | $java -jar /usr/bin/VarScan.v2.4.3.jar mpileup2cns $opts > ~/out/varscan_vcf/output/${bam_file_prefix[i]}.varscan.vcf
+		# Rename sample in vcf to corrospond to bam file name (aka sample name). Varscan defult is to name samples 'Sample1'
+		sed -i 's/Sample1/'"${bam_file_prefix[i]}"'/' ~/out/varscan_vcf/output/${bam_file_prefix[i]}.varscan.vcf
+		# if bedfile provided filter vcf to contain only variants within genomic regions specified by the bed file.
+		if [ "$bed_file" != "" ]; then
+			# use sed to remove chr from chromosome in bedfile. write to temp vcf (not output from this app)
+			sed 's/chr//' ~/out/varscan_vcf/output/${bam_file_prefix[i]}.varscan.vcf > ${bam_file_prefix[i]}.temp.vcf
+			# use bedtools to instersect BED file and VCF file. Write VCF direct to the output folder
+			/usr/bin/bedtools2/bin/bedtools intersect -header -a ${bam_file_prefix[i]}.temp.vcf -b ${bed_file_path} > ~/out/varscan_vcf_bed/output/${bam_file_prefix[i]}.varscan.bedfiltered.vcf
+		fi
 	fi
 fi
 done
